requirements.txt
pandas
mpld3
matplotlib
textblob
TextBlob
python_instagram